stat: Move all Tag constructors into lib/stat
