+++ node/lib/stat_emitter.js
@@ -22,7 +22,7 @@
 
 var inherits = require('util').inherits;
 var EventEmitter = require('./event_emitter');
-var Stat = require('./stat');
+var BaseStat = require('./stat').BaseStat;
 
 // This is an interface that exists only to explicitly declare the stats we are
 // tracking and avoid bugs due to mis-spelled names and types.
@@ -60,7 +60,7 @@ StatEmitter.prototype.emitStat = function emitStat(stat) {
 function emitStat(value, tags) {
     /*jshint validthis: true */
     var self = this;
-    var stat = new Stat(self.name, self.type, value, tags);
+    var stat = new BaseStat(self.name, self.type, value, tags);
     self.emitter.emitStat(stat);
 }
