+++ node/out_request.js
@@ -24,6 +24,7 @@ module.exports = TChannelOutRequest;
 
 var assert = require('assert');
 var EventEmitter = require('./lib/event_emitter');
+var stat = require('./lib/stat.js');
 var inherits = require('util').inherits;
 var parallel = require('run-parallel');
 
@@ -181,7 +182,7 @@ function emitResponseStat(res) {
             'outbound.calls.app-errors',
             'counter',
             1,
-            new OutboundCallsAppErrorsTags(
+            new stat.OutboundCallsAppErrorsTags(
                 self.serviceName,
                 self.headers.cn,
                 self.endpoint,
@@ -196,7 +197,7 @@ function emitOutboundCallsSuccess(request) {
         'outbound.calls.success',
         'counter',
         1,
-        new OutboundCallsSuccessTags(
+        new stat.OutboundCallsSuccessTags(
             request.serviceName,
             request.headers.cn,
             request.endpoint
@@ -204,32 +205,7 @@ function emitOutboundCallsSuccess(request) {
     ));
 }
 
-function OutboundCallsAppErrorsTags(serviceName, cn, endpoint, type) {
-    var self = this;
 
-    self.app = '';
-    self.host = '';
-    self.cluster = '';
-    self.version = '';
-
-    self.targetService = serviceName;
-    self.service = cn;
-    self.targetEndpoint = endpoint;
-    self.type = type;
-}
-
-function OutboundCallsSuccessTags(serviceName, cn, endpoint) {
-    var self = this;
-
-    self.app = '';
-    self.host = '';
-    self.cluster = '';
-    self.version = '';
-
-    self.targetService = serviceName;
-    self.service = cn;
-    self.targetEndpoint = endpoint;
-}
 
 TChannelOutRequest.prototype.emitPerAttemptResponseStat =
 function emitPerAttemptResponseStat(res) {
@@ -240,7 +216,7 @@ function emitPerAttemptResponseStat(res) {
             'outbound.calls.per-attempt.app-errors',
             'counter',
             1,
-            new OutboundCallsPerAttemptAppErrorsTags(
+            new stat.OutboundCallsPerAttemptAppErrorsTags(
                 self.serviceName,
                 self.headers.cn,
                 self.endpoint,
@@ -254,22 +230,6 @@ function emitPerAttemptResponseStat(res) {
     }
 };
 
-function OutboundCallsPerAttemptAppErrorsTags(
-    serviceName, cn, endpoint, type, retryCount
-) {
-    var self = this;
-
-    self.app = '';
-    self.host = '';
-    self.cluster = '';
-    self.version = '';
-
-    self.targetService = serviceName;
-    self.service = cn;
-    self.targetEndpoint = endpoint;
-    self.type = type;
-    self.retryCount = retryCount;
-}
 
 TChannelOutRequest.prototype.emitPerAttemptLatency =
 function emitPerAttemptLatency() {
@@ -281,7 +241,7 @@ function emitPerAttemptLatency() {
         'outbound.calls.per-attempt-latency',
         'timing',
         latency,
-        new OutboundCallsPerAttemptLatencyTags(
+        new stat.OutboundCallsPerAttemptLatencyTags(
             self.serviceName,
             self.headers.cn,
             self.endpoint,
@@ -290,23 +250,6 @@ function emitPerAttemptLatency() {
         )
     ));
 };
-
-function OutboundCallsPerAttemptLatencyTags(
-    serviceName, cn, endpoint, remoteAddr, retryCount
-) {
-    var self = this;
-
-    self.app = '';
-    self.host = '';
-    self.cluster = '';
-    self.version = '';
-
-    self.targetService = serviceName;
-    self.service = cn;
-    self.targetEndpoint = endpoint;
-    self.peer = remoteAddr;
-    self.retryCount = retryCount;
-}
 
 TChannelOutRequest.prototype.emitLatency = function emitLatency() {
     var self = this;
@@ -317,26 +260,13 @@ TChannelOutRequest.prototype.emitLatency = function emitLatency() {
         'outbound.calls.latency',
         'timing',
         latency,
-        new OutboundCallsLatencyTags(
+        new stat.OutboundCallsLatencyTags(
             self.serviceName,
             self.headers.cn,
             self.endpoint
         )
     ));
 };
-
-function OutboundCallsLatencyTags(serviceName, cn, endpoint) {
-    var self = this;
-
-    self.app = '';
-    self.host = '';
-    self.cluster = '';
-    self.version = '';
-
-    self.targetService = serviceName;
-    self.service = cn;
-    self.targetEndpoint = endpoint;
-}
 
 TChannelOutRequest.prototype.emitError = function emitError(err) {
     var self = this;
@@ -477,26 +407,13 @@ function emitOutboundCallsSent() {
         'outbound.calls.sent',
         'counter',
         1,
-        new OutboundCallsSentTags(
+        new stat.OutboundCallsSentTags(
             self.serviceName,
             self.headers.cn,
             self.endpoint
         )
     ));
 };
-
-function OutboundCallsSentTags(serviceName, cn, endpoint) {
-    var self = this;
-
-    self.app = '';
-    self.host = '';
-    self.cluster = '';
-    self.version = '';
-
-    self.targetService = serviceName;
-    self.service = cn;
-    self.targetEndpoint = endpoint;
-}
 
 TChannelOutRequest.prototype.hookupStreamCallback =
 function hookupCallback(callback) {
