+++ node/v2/handler.js
@@ -21,6 +21,7 @@
 'use strict';
 
 var EventEmitter = require('../lib/event_emitter');
+var stat = require('../lib/stat.js');
 var util = require('util');
 var assert = require('assert');
 
@@ -205,7 +206,7 @@ TChannelV2Handler.prototype.handleCallRequest = function handleCallRequest(reqFr
         'inbound.request.size',
         'counter',
         reqFrame.size,
-        new InboundRequestSizeTags(
+        new stat.InboundRequestSizeTags(
             req.headers.cn,
             req.serviceName,
             req.endpoint
@@ -225,38 +226,13 @@ function emitBytesRecvd(frame) {
             'connections.bytes-recvd',
             'counter',
             frame.size,
-            new ConnectionsBytesRcvdTags(
+            new stat.ConnectionsBytesRcvdTags(
                 channel.hostPort || '0.0.0.0:0',
                 self.connection.socketRemoteAddr
             )
         ));
     }
 };
-
-function InboundRequestSizeTags(cn, serviceName, endpoint) {
-    var self = this;
-
-    self.app = null;
-    self.host = null;
-    self.cluster = null;
-    self.version = null;
-
-    self.callingService = cn;
-    self.service = serviceName;
-    self.endpoint = endpoint;
-}
-
-function ConnectionsBytesRcvdTags(hostPort, peerHostPort) {
-    var self = this;
-
-    self.app = null;
-    self.host = null;
-    self.cluster = null;
-    self.version = null;
-
-    self.hostPort = hostPort;
-    self.peerHostPort = peerHostPort;
-}
 
 TChannelV2Handler.prototype.incomingRequestInvalid =
 function incomingRequestInvalid(reqFrame, req) {
@@ -335,7 +311,7 @@ TChannelV2Handler.prototype.handleCallResponse = function handleCallResponse(res
         'inbound.response.size',
         'counter',
         resFrame.size,
-        new InboundResponseSizeTags(
+        new stat.InboundResponseSizeTags(
             req ? req.headers.cn : '',
             req ? req.serviceName : '',
             req ? req.endpoint : ''
@@ -354,18 +330,7 @@ TChannelV2Handler.prototype.handleCallResponse = function handleCallResponse(res
     }
 };
 
-function InboundResponseSizeTags(cn, serviceName, endpoint) {
-    var self = this;
 
-    self.app = null;
-    self.host = null;
-    self.cluster = null;
-    self.version = null;
-
-    self.callingService = cn;
-    self.service = serviceName;
-    self.endpoint = endpoint;
-}
 
 TChannelV2Handler.prototype.checkValidCallResponse =
 function checkValidCallResponse(resFrame) {
@@ -427,7 +392,7 @@ TChannelV2Handler.prototype.handleCallRequestCont = function handleCallRequestCo
         'inbound.request.size',
         'counter',
         reqFrame.size,
-        new InboundRequestSizeTags(
+        new stat.InboundRequestSizeTags(
             req.headers.cn,
             req.serviceName,
             req.endpoint
@@ -455,7 +420,7 @@ TChannelV2Handler.prototype.handleCallResponseCont = function handleCallResponse
         'inbound.response.size',
         'counter',
         resFrame.size,
-        new InboundResponseSizeTags(
+        new stat.InboundResponseSizeTags(
             req ? req.headers.cn : '',
             req ? req.serviceName : '',
             req ? req.endpoint : ''
@@ -591,7 +556,7 @@ function sendCallRequestFrame(req, flags, args) {
         'outbound.request.size',
         'counter',
         result.size,
-        new OutboundRequestSizeTags(
+        new stat.OutboundRequestSizeTags(
             req.serviceName,
             req.headers.cn,
             req.endpoint
@@ -611,7 +576,7 @@ function emitBytesSent(result) {
             'connections.bytes-sent',
             'counter',
             result.size,
-            new ConnectionsBytesSentTags(
+            new stat.ConnectionsBytesSentTags(
                 channel.hostPort || '0.0.0.0:0',
                 self.connection.socketRemoteAddr
             )
@@ -658,31 +623,6 @@ function verifyCallRequestFrame(req, args) {
     return true;
 };
 
-function OutboundRequestSizeTags(serviceName, cn, endpoint) {
-    var self = this;
-
-    self.app = null;
-    self.host = null;
-    self.cluster = null;
-    self.version = null;
-
-    self.targetService = serviceName;
-    self.service = cn;
-    self.targetEndpoint = endpoint;
-}
-
-function ConnectionsBytesSentTags(hostPort, peer) {
-    var self = this;
-
-    self.app = null;
-    self.host = null;
-    self.cluster = null;
-    self.version = null;
-
-    self.hostPort = hostPort;
-    self.peerHostPort = peer;
-}
-
 TChannelV2Handler.prototype.sendCallResponseFrame = function sendCallResponseFrame(res, flags, args) {
     var self = this;
     if (self.remoteName === null) {
@@ -707,7 +647,7 @@ TChannelV2Handler.prototype.sendCallResponseFrame = function sendCallResponseFra
         'outbound.response.size',
         'counter',
         result.size,
-        new OutboundResponseSizeTags(
+        new stat.OutboundResponseSizeTags(
             req.serviceName,
             req.headers.cn,
             req.endpoint
@@ -716,19 +656,6 @@ TChannelV2Handler.prototype.sendCallResponseFrame = function sendCallResponseFra
 
     self.emitBytesSent(result);
 };
-
-function OutboundResponseSizeTags(serviceName, cn, endpoint) {
-    var self = this;
-
-    self.app = null;
-    self.host = null;
-    self.cluster = null;
-    self.version = null;
-
-    self.targetService = serviceName;
-    self.service = cn;
-    self.targetEndpoint = endpoint;
-}
 
 TChannelV2Handler.prototype.validateCallResponseFrame =
 function validateCallResponseFrame(res) {
@@ -764,7 +691,7 @@ TChannelV2Handler.prototype.sendCallRequestContFrame = function sendCallRequestC
         'outbound.request.size',
         'counter',
         result.size,
-        new OutboundRequestSizeTags(
+        new stat.OutboundRequestSizeTags(
             req0 ? req0.serviceName : '',
             req0 ? req0.headers.cn : '',
             req0 ? req.endpoint : ''
@@ -791,7 +718,7 @@ TChannelV2Handler.prototype.sendCallResponseContFrame = function sendCallRespons
         'outbound.response.size',
         'counter',
         result.size,
-        new OutboundResponseSizeTags(
+        new stat.OutboundResponseSizeTags(
             req.serviceName,
             req.headers.cn,
             req.endpoint
